Move MR classes out of API into hadoopImpl
