+++ hadoop-mapreduce/src/main/java/org/apache/accumulo/hadoopImpl/mapreduce/lib/partition/RangePartitioner.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.accumulo.hadoop.mapreduce.lib.partition;
+package org.apache.accumulo.hadoopImpl.mapreduce.lib.partition;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
