+++ hadoop-mapreduce/src/test/java/org/apache/accumulo/hadoop/mapred/AccumuloMultiTableInputFormatTest.java
@@ -26,7 +26,7 @@ import java.util.Map;
 import org.apache.accumulo.core.client.IteratorSetting;
 import org.apache.accumulo.core.data.Range;
 import org.apache.accumulo.core.util.Pair;
-import org.apache.accumulo.hadoop.mapreduce.InputTableConfig;
+import org.apache.accumulo.hadoopImpl.mapreduce.InputTableConfig;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapred.JobConf;
 import org.junit.Rule;
@@ -39,7 +39,7 @@ public class AccumuloMultiTableInputFormatTest {
   public TestName testName = new TestName();
 
   /**
-   * Verify {@link org.apache.accumulo.hadoop.mapreduce.InputTableConfig} objects get correctly
+   * Verify {@link InputTableConfig} objects get correctly
    * serialized in the JobContext.
    */
   @Test
